boss' abilities, cleaning up the scripts, limit of spawns
Bullet Script

-removed the redundant public GameObject target declaration and made it a private field
-renamed the _damage and _moveDirection fields to be private and follow the naming convention of private fields
-created two public methods, setBulletDamage and setMovementDirection, to set the bullet's damage and movement direction, respectively
-in the Start method, updated the code to check for the presence of the target before proceeding to calculate the bullet's velocity based on the target's position
-adjusted the code to use the _damage field when dealing damage to the player

EnemySpawner Script

-inttroduced a new private field enemiesSpawned to track the number of enemies spawned by the spawner
-updated the FixedUpdate method to check if the number of enemies spawned (enemiesSpawned) is less than 3 before instantiating a new enemy;this ensures that no more than three enemies are spawned by the spawner
-icremented the enemiesSpawned count each time an enemy is instantiated to keep track of the total number spawned

EnemyMovement Script

-enhanced the waypoint handling logic to allow for more precise movement and waypoint navigation
-improved the firing logic by introducing two fire rates: FireRateFireLotsOfBullets and FireRateFireBullet
-separated the logic for firing a single bullet and firing a lot of bullets
-added cooldown periods for both types of firing
-ensured that the enemy fires bullets only when the player is within the specified shooting range

+++ Assets/Scripts/player/PlayerHealth.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 
 public class PlayerHealth : MonoBehaviour
 {
-    [SerializeField] private int maxHealth = 10;
+    [SerializeField] private float maxHealth = 10;
     public Health healthBar;
 
-    private int health;
+    private float health;
 
     // Start is called before the first frame update
     void Start()
@@ -18,7 +18,7 @@ public class PlayerHealth : MonoBehaviour
         healthBar.SetMaxHealth(maxHealth); 
     }
     //increases the player's health 
-    public void AddHealth(int amount)
+    public void AddHealth(float amount)
     {
         health += amount;
         health = Mathf.Min(health, maxHealth);
@@ -26,7 +26,7 @@ public class PlayerHealth : MonoBehaviour
     }
 
     //inflicts damage to the player's health
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         health -= damage;
